+++ src/main/java/frc/robot/subClass/Arm.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.subClass;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.DemandType;
@@ -13,14 +13,14 @@ import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 // import edu.wpi.first.wpilibj.XboxController;
 /**import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard; **/
-
+//import frc.robot.subClass.State.*;
 
 public class Arm {
     private SensorCollection armPoint;
     private WPI_TalonSRX armMotor;
     
 
-    Arm(){
+    public Arm(){
         armMotor = new WPI_TalonSRX(Const.ArmMotor);
         armPoint = new SensorCollection(armMotor); 
         armMotor.configSelectedFeedbackSensor(FeedbackDevice.Analog, Const.SFSPidIdx, Const.SFSTimeoutMS);
@@ -39,6 +39,7 @@ public class Arm {
         double pointRange = Const.HighPoint - Const.LowPoint;
         return angleDiff*(pointRange/angleRange) + Const.LowPoint;
     }
+    
     public double pointToAngle(double Point) {
         double pointDiff = Point - Const.LowPoint;
         double angleRange = Const.HighAngle - Const.LowAngle;
@@ -46,16 +47,20 @@ public class Arm {
         return pointDiff*(angleRange/pointRange) + Const.LowAngle;
         
     }
+    
     public double getArmAngle() {
         return pointToAngle(armPoint.getAnalogInRaw());
     }
+   
     public void armSet(ControlMode controlMode, double input) {
         armMotor.set(controlMode, input);
     }
+   
     public void ArmPIDMove(double Angle) {
         armMotor.set(ControlMode.Position, angleToPoint(Angle),
         DemandType.ArbitraryFeedForward, Const.ArmGCoef*Math.cos(Math.toRadians(getArmAngle())));
     }
+   
     public void ArmRelease(){
         armMotor.set(ControlMode.PercentOutput, Const.ArmFFCoef*Math.cos(Math.toRadians(getArmAngle())));
     }
