+++ src/main/java/frc/robot/subClass/Const.java
@@ -1,4 +1,5 @@
-package frc.robot;
+package frc.robot.subClass;
+
 
 public class Const {
     // motor 
@@ -9,6 +10,7 @@ public class Const {
     public static final int DriveController = 0;
     public static final int IntakeMotor = 14;
     public static final int ArmMotor = 3;
+
     // Deadband and intakemotor
     public static final double Deadband = 0.03;
     public static final double IntakeSpeed = -0.3;
@@ -16,14 +18,17 @@ public class Const {
     public static final double IntakeNeutral = 0;
     public static final double ShooterLeftOut = -0.7;
     public static final double ShooterRightOut = 0.7;
+
     // angle to point, point to angle
     public static final int LowAngle = -32;
     public static final int HighAngle = 86;
     public static final int LowPoint = 181;
     public static final int HighPoint = 516;
+
     // Arm, Anti-gravity (Coef = coeffiecient, FF = freefall, G = gravity)
     public static final double ArmGCoef = 0.13;
     public static final double ArmFFCoef = -0.04;
+
     // PDIF (V = value)
     public static final int FslotIdx = 0;
     public static final double FVofP = 0;
@@ -47,6 +52,7 @@ public class Const {
 
     // feedback sensor
     public static final int SFSPidIdx = 0;
+
     // 0 又は 1
     public static final int SFSTimeoutMS = 30;
